requirements.txt
python-dotenv
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
pytz
dateparser